+++ openyich-framework/src/main/java/com/openyich/framework/boot/autoconfigure/OpenYichProperties.java
@@ -346,6 +346,8 @@ public class OpenYichProperties {
     private String[] protocols = {};
 
     private boolean useDefaultResponseMessages = true;
+    
+    private final Authorization authorization = new Authorization();;
 
     public String getTitle() {
       return title;
@@ -449,6 +451,84 @@ public class OpenYichProperties {
 
     public void setUseDefaultResponseMessages(final boolean useDefaultResponseMessages) {
       this.useDefaultResponseMessages = useDefaultResponseMessages;
+    }
+    
+    public Authorization getAuthorization() {
+      return authorization;
+    }
+
+    public static class Authorization {
+      
+      private boolean enabled;
+      
+      private String name = "Authorization";
+      
+      private String description = "Add authorization header";
+      
+      private String defaultValue = "Bearer ";
+      
+      private boolean required;
+      
+      private String paramType = "header";
+      
+      private String pattern = "";
+
+      public boolean isEnabled() {
+        return enabled;
+      }
+
+      public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+      }
+
+      public String getName() {
+        return name;
+      }
+
+      public void setName(String name) {
+        this.name = name;
+      }
+
+      public String getDescription() {
+        return description;
+      }
+
+      public void setDescription(String description) {
+        this.description = description;
+      }
+
+      public String getDefaultValue() {
+        return defaultValue;
+      }
+
+      public void setDefaultValue(String defaultValue) {
+        this.defaultValue = defaultValue;
+      }
+
+      public boolean isRequired() {
+        return required;
+      }
+
+      public void setRequired(boolean required) {
+        this.required = required;
+      }
+
+      public String getParamType() {
+        return paramType;
+      }
+
+      public void setParamType(String paramType) {
+        this.paramType = paramType;
+      }
+
+      public String getPattern() {
+        return pattern;
+      }
+
+      public void setPattern(String pattern) {
+        this.pattern = pattern;
+      }
+      
     }
   }